Chore :: pull을 하기 위한 빌드업
